requirements.txt
APScheduler==3.6.3
astroid==2.4.2
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mergedeep==1.3.0
pycryptodome==3.10.1
pylint==2.6.0
pymongo==3.11.2
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
toml==0.10.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.2
wrapt==1.12.1
